Consolidate all evm balance checks into one file
