+++ packages/token-balance-cache/src/util/validateOpts.ts
@@ -2,14 +2,10 @@ import type { EthBPOpts } from '../types';
 import Web3 from 'web3';
 
 export function validateOpts(
-  name: string,
   address: string,
   opts: EthBPOpts & { tokenId?: string }
 ) {
   const { tokenAddress, contractType, tokenId } = opts;
-  if (contractType != name) {
-    throw new Error('Invalid Contract Type');
-  }
   if (!tokenAddress) {
     throw new Error('Need Token Address');
   }
